sudoers: Reset env, set secure path
I'm not quite sure why it only happens on Lin's laptop, but without
this, $PATH is polluted on the other side of sudo. This usually doesn't
matter, but plenv won't update the shim location on initialisation with
it there, which results in having $PLENV_ROOT set to /home/lin/.plenv
(as it's set in every plenv shim), and thus failing to create
directories needed as we're another user.